showing integration icon
